+++ src/primitives/index.js
@@ -5,3 +5,4 @@ export Header from './primitive-header.component';
 export Title from './primitive-title.component';
 export Subtitle from './primitive-subtitle.component';
 export Block from './primitive-block.component';
+export Spinner from './primitive-spinner.component';